+++ src/main/java/com/chauke/clinicdatabase/repository/EmployeeRepository.java
@@ -3,5 +3,5 @@ package com.chauke.clinicdatabase.repository;
 import com.chauke.clinicdatabase.entity.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployeeRepository extends JpaRepository<Employee, Long> {
+public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
 }
